+++ zone/lua_general.cpp
@@ -200,7 +200,7 @@ void unregister_player_event(int evt) {
 
 void register_item_event(std::string name, int evt, int item_id, luabind::adl::object func) {
 	std::string package_name = "item_";
-	package_name += std::to_string(static_cast<long long>(item_id));
+	package_name += std::to_string(item_id);
 
 	if(luabind::type(func) == LUA_TFUNCTION) {
 		register_event(package_name, name, evt, func);
@@ -214,7 +214,7 @@ void register_item_event(int evt, int item_id, luabind::adl::object func) {
 
 void unregister_item_event(std::string name, int evt, int item_id) {
 	std::string package_name = "item_";
-	package_name += std::to_string(static_cast<long long>(item_id));
+	package_name += std::to_string(item_id);
 
 	unregister_event(package_name, name, evt);
 }
